requirements.txt
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.19.1
email-validator==2.0.0.post2
executing==2.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.20.0
filelock==3.12.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.54.1
fqdn==1.5.1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
idna==3.4
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.1
orjson==3.10.18
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
Pillow==10.0.1
platformdirs==3.10.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.20.0
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.20
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20240906
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.24.3
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
WTForms==3.0.1
